[REF] web: move legacy dom helpers to public
as it is only used/necessary there (not used anymore in the
backend).

Part of task 3586757
