+++ addons/web/__manifest__.py
@@ -154,7 +154,7 @@ This module provides the core of the Odoo Web Client.
             'web/static/src/session.js',
             'web/static/src/core/browser/cookie.js',
             'web/static/src/core/utils/ui.js',
-            'web/static/src/legacy/js/core/minimal_dom.js',
+            'web/static/src/legacy/js/public/minimal_dom.js',
             'web/static/src/legacy/js/public/lazyloader.js',
         ],
         'web.assets_frontend': [
@@ -259,7 +259,7 @@ This module provides the core of the Odoo Web Client.
             ('remove', 'web/static/src/session.js'),
             ('remove', 'web/static/src/core/browser/cookie.js'),
             ('remove', 'web/static/src/core/utils/ui.js'),
-            ('remove', 'web/static/src/legacy/js/core/minimal_dom.js'),
+            ('remove', 'web/static/src/legacy/js/public/minimal_dom.js'),
             ('remove', 'web/static/src/legacy/js/public/lazyloader.js'),
         ],
         'web.report_assets_common': [
